app/requirements.txt
flask
scikit-learn
pandas
numpy
joblib
gunicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask